+++ src/lib/filter/filterRulesWithLogs.ts
